+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.2 - 2016-03-06
+
+* __Adds__ function wrapper that returns a `new OError` instance. To use:
+
+	```javascript
+	var oerror = require('oerror').oerror;
+	var err = oerror({ /* OError constructor options */ });
+	```
+
 # v0.1 - 2016-03-01
 
 Initial release.